update bridge to match agent
